updates during balance testing

+++ src/main/java/frc/robot/subsystem/DriveSubsystem.java
@@ -324,7 +324,7 @@ public class DriveSubsystem extends SubsystemBase {
   /** 
    * @return double
    */
-  public double getPitch() {
+  public int getPitch() {
     return m_pigeon.getPitch();
   }
